feat!: tweak tsconfig
We don't need DOM types -- these packages should work without a browser.
We can target ES2022 because Node.js 18 (our min) supports it, and
using this in an older browser can be facilitated with a bundler
We can use module node18 because our min TypeScript version is 5.9.2,
and 5.8^ supports node18
We use node16 module resolution because it's not designed to track node
versions, it only increments when it changes. nodenext is unpinned
whilst node16 is a fixed spec.